Fix table reference in field type instances
Fixed the table reference of the settings table in the field type instances.
Now it is possible to have an individual configuration for each instance of a field type which is very important for multiple foreign keys in the same table.
This fix requires the settings table to be a one dimensional table (which it currently is).